spike: decompose payload construction
