+++ crates/builder/core/src/flashblocks/mod.rs
@@ -1,16 +1,6 @@
-//! Flashblocks builder types.
-
-mod best_txs;
-pub use best_txs::BestFlashblocksTxs;
-
-mod cell;
-pub use cell::{BlockCell, WaitForValue};
-
+/// Flashblocks builder types.
 mod resolve;
 pub use resolve::ResolvePayload;
-
-mod args;
-pub use args::BuildArguments;
 
 mod job;
 pub use job::BlockPayloadJob;
@@ -18,23 +8,8 @@ pub use job::BlockPayloadJob;
 mod generator;
 pub use generator::BlockPayloadJobGenerator;
 
-mod traits;
-pub use traits::PayloadBuilder;
-
 mod handler;
 pub use handler::PayloadHandler;
-
-mod config;
-pub use config::FlashblocksConfig;
-
-mod context;
-pub use context::{FlashblocksExtraCtx, OpPayloadBuilderCtx};
-
-mod info;
-pub use info::FlashblocksExecutionInfo;
-
-mod payload;
-pub use payload::OpPayloadBuilder;
 
 mod service;
 pub use service::FlashblocksServiceBuilder;
