--- crates/builder/core/src/flashblocks/traits.rs
@@ -1,38 +1,0 @@
-//! Contains the payload builder trait.
-
-use reth_node_api::PayloadBuilderAttributes;
-use reth_payload_builder::PayloadBuilderError;
-use reth_payload_primitives::BuiltPayload;
-
-use crate::{BlockCell, BuildArguments};
-
-/// A trait for building payloads that encapsulate Ethereum transactions.
-///
-/// This trait provides the `try_build` method to construct a transaction payload
-/// using `BuildArguments`. It returns a `Result` indicating success or a
-/// `PayloadBuilderError` if building fails.
-#[async_trait::async_trait]
-pub trait PayloadBuilder: Send + Sync + Clone {
-    /// The payload attributes type to accept for building.
-    type Attributes: PayloadBuilderAttributes;
-    /// The type of the built payload.
-    type BuiltPayload: BuiltPayload;
-
-    /// Tries to build a transaction payload using provided arguments.
-    ///
-    /// Constructs a transaction payload based on the given arguments,
-    /// returning a `Result` indicating success or an error if building fails.
-    ///
-    /// # Arguments
-    ///
-    /// - `args`: Build arguments containing necessary components.
-    ///
-    /// # Returns
-    ///
-    /// A `Result` indicating the build outcome or an error.
-    async fn try_build(
-        &self,
-        args: BuildArguments<Self::Attributes, Self::BuiltPayload>,
-        best_payload: BlockCell<Self::BuiltPayload>,
-    ) -> Result<(), PayloadBuilderError>;
-}
